never support decoding into an already-populated oneof
this is never needed because we don't support oneof values encoded in multiple fields, and oneofs are a class of cases where the presence of a value is always fully unambiguous so we don't need "duplicated" either.

we also always add field name diagnostics to the error path here inside oneofs, formerly only done in messages.